+++ tests/datasets/tests/test_numerical.py
@@ -9,7 +9,7 @@ class TestRandomIntegerGenerator:
     def test(self):
         output = numerical.RandomIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) > 1
         assert np.isnan(output).sum() == 0
 
@@ -19,7 +19,7 @@ class TestRandomIntegerNaNsGenerator:
     def test(self):
         output = numerical.RandomIntegerNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) > 1
         assert np.isnan(output).sum() > 0
 
@@ -29,7 +29,7 @@ class TestConstantIntegerGenerator:
     def test(self):
         output = numerical.ConstantIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) == 1
         assert np.isnan(output).sum() == 0
 
@@ -39,7 +39,7 @@ class TestConstantIntegerNaNsGenerator:
     def test(self):
         output = numerical.ConstantIntegerNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 2
         assert np.isnan(output).sum() >= 1
 
@@ -49,7 +49,7 @@ class TestAlmostConstantIntegerGenerator:
     def test(self):
         output = numerical.AlmostConstantIntegerGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.int
+        assert output.dtype == int
         assert len(pd.unique(output)) == 2
         assert np.isnan(output).sum() == 0
 
@@ -59,7 +59,7 @@ class TestAlmostConstantIntegerNaNsGenerator:
     def test(self):
         output = numerical.AlmostConstantIntegerNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 3
         assert np.isnan(output).sum() >= 1
 
@@ -69,7 +69,7 @@ class TestNormalGenerator:
     def test(self):
         output = numerical.NormalGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 10
         assert np.isnan(output).sum() == 0
 
@@ -79,7 +79,7 @@ class TestNormalNaNsGenerator:
     def test(self):
         output = numerical.NormalNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert 1 < len(pd.unique(output)) <= 10
         assert np.isnan(output).sum() >= 1
 
@@ -89,7 +89,7 @@ class TestBigNormalGenerator:
     def test(self):
         output = numerical.BigNormalGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert len(pd.unique(output)) == 10
         assert np.isnan(output).sum() == 0
 
@@ -99,6 +99,6 @@ class TestBigNormalNaNsGenerator:
     def test(self):
         output = numerical.BigNormalNaNsGenerator.generate(10)
         assert len(output) == 10
-        assert output.dtype == np.float
+        assert output.dtype == float
         assert 1 < len(pd.unique(output)) <= 10
         assert np.isnan(output).sum() >= 1
